refactor: move hooks and helpers out of components/organisms to separate concerns
