+++ src/lib/workbench-helpers.ts
@@ -1,4 +1,4 @@
-import type { PromptSegment } from "../../shared/lib/db-schema"
+import type { PromptSegment } from "../shared/lib/db-schema"
 
 export async function sendToWorkbench(
   value: string,
